+++ UrlStatus.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HttpsSpider", "WindowsFormsApplication1\HttpsSpider.csproj", "{F987D9A8-5C41-4032-8AD3-EC06205957B1}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UrlStatus", "UrlStatus\UrlStatus.csproj", "{F987D9A8-5C41-4032-8AD3-EC06205957B1}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C433B380-57CC-4F31-8ABB-A59E1D29120A}"
 	ProjectSection(SolutionItems) = preProject
