+++ UrlStatus/Main.cs
@@ -8,7 +8,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace WindowsFormsApplication1
+namespace UrlStatus
 {
     public partial class Main : Form
     {
