+++ UrlStatus/UrlStatus.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{F987D9A8-5C41-4032-8AD3-EC06205957B1}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>WindowsFormsApplication1</RootNamespace>
-    <AssemblyName>WindowsFormsApplication1</AssemblyName>
+    <RootNamespace>UrlStatus</RootNamespace>
+    <AssemblyName>UrlStatus</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
