--- WindowsFormsApplication1/log4net.config
@@ -1,34 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-
-  <log4net>
-    <root>
-      <!--level: OFF>FATAL>ERROR>WARN>INFO>DEBUG>ALL 缺省值为DEBUG-->
-      <level value="ALL"/>
-      <appender-ref ref="RollingFileAppender" />
-    </root>
-    <logger name="FileLogger">
-      <level value="DEBUG" />
-      <appender-ref ref="RollingFileAppender" />
-    </logger>
-    <appender name="RollingFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file value="logs/log-" />
-      <appendToFile value="true" />
-      <rollingStyle value="Composite" />
-      <lockingModel type="log4net.Appender.FileAppender+MinimalLock" />
-      <maxSizeRollBackups value="-1" />
-      <maximumFileSize value="2MB" />
-      <!--<datePattern value="yyyy/yyyyMM/'APP-'yyyyMMdd'.log'" />-->
-      <datePattern value="'APP-'yyyyMMdd'.log'" />
-      <staticLogFileName value="false"/>
-      <encoding value="utf-8"/>
-      <layout type="log4net.Layout.PatternLayout">
-        <!--<conversionPattern value="%d [%t] %-5p %logger [%p{NDC}] - %m%n" />-->
-        <!--<param name="ConversionPattern" value="记录时间：%date 线程ID:[%thread] 日志级别：%-5level 记录类：%logger 操作者ID：%property{Operator} 操作类型：%property{Action}%n 当前机器名:%property%n当前机器名及登录用户：%username %n 记录位置：%location%n 消息描述：%property{Message}%n 异常：%exception%n 消息：%message%newline%n%n" />-->
-        <param name="ConversionPattern" value="%date [%thread] %-5level%n %location%n %message%n %exception%n" />
-      </layout>
-    </appender>
-  </log4net>
-  <!--private static log4net.ILog log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);-->
-
-</configuration>
